==> 1_merge_commits_coletados.ipynb <==
# Setup

import datetime
# !pip install pytz
import pytz
import pandas as pd

# 1. Leitura de arquivos

df1 = pd.read_excel('/content/commits_projetos_1.xlsx')
df2 = pd.read_excel('/content/commits_projetos_2.xlsx')
df3 = pd.read_excel('/content/commits_projetos_3.xlsx')
df4 = pd.read_csv('/content/commits_projetos_4.csv')
df5 = pd.read_csv('/content/commits_projetos_5.csv')

df1.shape, df2.shape, df3.shape, df4.shape, df5.shape

df1.head(1)

df2.head(1)

df3.head(1)

df4.head(1)

df5.head(1)

# 2. Merge de dados

Juntando todos os dados que temos de commits coletados

all_data = pd.concat([df1, df2, df3, df4, df5])
all_data.shape

all_data.shape

all_data.head(1)

Removendo commits duplicados

# considerando apenas 'sha' exclui-se repos 'forkados'
merged_data_df = all_data.drop_duplicates(subset=['sha', 'repo'], keep="last")
merged_data_df.shape

merged_data_df.isna().sum()

# 3. Salvar merge de commits

brasilia_tz = pytz.timezone('America/Sao_Paulo')
now = datetime.datetime.now(brasilia_tz).strftime("%d-%m-%Y_%H-%M-%S")
print(now)

merged_data_df.to_csv(f"merged_commits_{now}.csv", index=False)

all_data.to_csv(f"all_commits_{now}.csv", index=False)
