==> 2_get_repositories_info.ipynb <==
# Setup

from ipywidgets import FileUpload, interact

import requests
import pandas as pd
import time

# Autenticação API Github

@interact(files=FileUpload())
def set_token(files={}):
    global GITHUB_TOKEN
    if files:
        if isinstance(files, tuple): # Jupyter Notebook
            file = files[0]
            token = file['content'].tobytes().decode("utf-8").strip()
        elif isinstance(files, dict): # Google Colab
            file = list(files.values())[0]
            GITHUB_TOKEN = file['content'].decode("utf-8").strip()
        print("Token Loaded!")

SITE = "https://api.github.com"
def token_auth(request):
    request.headers["User-Agent"] = "MetodosPesquisa" # Required
    request.headers["Authorization"] = "token {}".format(GITHUB_TOKEN)
    return request
response = requests.get(SITE, auth=token_auth)
response.status_code

github_endpoints = response.json()
github_endpoints

# Obter nomes de repositórios através dos commits coletados anteriormente

collected_commits_df = pd.read_csv('/content/merged_commits_11-11-2025_12-57-06.csv')
print(collected_commits_df.shape)
collected_commits_df.head()

repos = collected_commits_df['repo'].unique().tolist()
len(repos)

# Coletar dados dos repositórios listados

# collected_commits_df_old = pd.read_excel('/content/commits_projetos.xlsx')
# repos_old = collected_commits_df_old['repo'].unique().tolist()
# len(repos_old)

def get_all_keys(d):
    keys = []
    for key, value in d.items():
        keys.append(key)
        if isinstance(value, dict):
            keys.extend(get_all_keys(value))
    return keys


def flatten_dict(nested_dict, parent_key='', sep='/') -> dict:
    items = []
    for key, value in nested_dict.items():
        new_key = f"{parent_key}{sep}{key}" if parent_key else key

        if isinstance(value, dict):
            items.extend(flatten_dict(value, new_key, sep=sep).items())
        else:
            items.append((new_key, value))

    return dict(items)

all_repos_info = []

for repo in repos:
    owner_name, repo_name = repo.split("/")
    # print(owner_name, repo_name)
    repo_response_raw = requests.get(github_endpoints['repository_url'].format(owner=owner_name, repo=repo_name), auth=token_auth)
    # print(repo_response_raw.status_code)
    repo_response = repo_response_raw.json()

    all_keys = get_all_keys(repo_response)
    repo_info = flatten_dict(repo_response)
    repo_info["all_keys"] = all_keys
    all_repos_info.append(repo_info)

repo_df = pd.DataFrame(all_repos_info)
repo_df.shape

**Obs.:** Às vezes retorna 432 colunas, outras vezes retorna 238.

repo_df.head(1)

repo_df.shape

new = repo_df.copy()
new.shape

sos = repo_df.copy()
sos.shape

# Salvar dados coletados

print(repo_df.shape)
repo_df.to_csv("repositories_info.csv", index=False)

# Pré-processamento dos dados dos repositórios

## Verificando retornos vazios

repo_df["full_name"].isna().sum()

[lost_repo for lost_repo in repos if lost_repo not in repo_df["full_name"].values.tolist()]

**3 repositórios não estão mais disponíveis**. Eles podem ter sido removidos ou sofreram restrições de acesso público.

print(repo_df.shape)
repo_df.dropna(subset="full_name", inplace=True)
print(repo_df.shape)

## Remoção de colunas indesejadas

repo_df.columns

owner_cols = [col for col in repo_df.columns.tolist() if col.startswith("owner")] # owner/type
license_cols = [col for col in repo_df.columns.tolist() if col.startswith("license")]
template_cols = [col for col in repo_df.columns.tolist() if col.startswith("template_repository")]
permission_cols = [col for col in repo_df.columns.tolist() if col.startswith("permissions")]

cols_to_drop = [
    'id', 'node_id', 'private', 'name',
    'keys_url', 'teams_url', 'hooks_url', 'blobs_url' , 'git_tags_url', 'git_refs_url' , 'trees_url',
    'statuses_url', 'collaborators_url', 'issue_events_url',
    'git_url', 'ssh_url', 'clone_url', 'svn_url', 'homepage', 'watchers_count', 'license',
    'allow_forking', 'is_template', 'visibility',
    'commits_url', 'git_commits_url', 'comments_url', 'issue_comment_url', 'compare_url',
    'contents_url', #
    'archive_url', 'issues_url', 'milestones_url', 'releases_url', 'labels_url', 'notifications_url',
    'pulls_url',
    'all_keys' # coluna de segurança, deixar apenas na versão full



]
cols_to_drop.extend(owner_cols)
cols_to_drop.extend(license_cols)
cols_to_drop.extend(template_cols)
cols_to_drop.extend(permission_cols)

repo_df.shape, len(cols_to_drop)

repo_df_filtered = repo_df.drop(columns=cols_to_drop)
repo_df.shape, repo_df_filtered.shape

repo_df_filtered.to_csv("repositories_info_reduced.csv", index=False)

repo_df_filtered.isna().sum()

## Separando dados de fork

repo_df_filtered["fork"].value_counts()

fork_df = repo_df_filtered[repo_df_filtered["fork"]].copy()
fork_df.shape

fork_df

common_df = repo_df_filtered[repo_df_filtered["fork"] == False].copy()
common_df.shape

## Limpando dados de repositórios comuns (sem fork)

repo_df_filtered["parent/id"].value_counts()

repo_df_filtered["source/id"].value_counts()

fork_cols = [col for col in repo_df.columns.tolist() if col.startswith("source") or col.startswith("parent")]
fork_cols.extend(["forks_url"])

common_df_filtered = common_df.drop(columns=fork_cols)
common_df.shape, common_df_filtered.shape


## Salvando dados de repositórios após ajustes

common_df_filtered.head(1)

fork_df.to_csv("forked_repos.csv", index=False)
# common_df_filtered.to_csv("repositories_info_reduced_no_fork.csv", index=False)
common_df_filtered.to_csv("common_repos.csv", index=False)
